decisionNeuralNetwork/requirements_minimal.txt
# Core PyTorch and ML dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pillow>=8.0.0

# Data science and visualization
matplotlib>=3.5.0
scikit-learn>=1.0.0

# Additional utilities
tqdm>=4.60.0
